muitas modificações, email, recuperação de senhas, aproveitamento de estágio, drive, entre outras

+++ src/main/java/br/edu/ifrs/restinga/assinaturadigitalestagioifrsrestingaapi/controller/ServidorController.java
@@ -8,18 +8,17 @@ import br.edu.ifrs.restinga.assinaturadigitalestagioifrsrestingaapi.infra.error.
 import br.edu.ifrs.restinga.assinaturadigitalestagioifrsrestingaapi.infra.security.TokenService;
 import br.edu.ifrs.restinga.assinaturadigitalestagioifrsrestingaapi.model.Curso;
 import br.edu.ifrs.restinga.assinaturadigitalestagioifrsrestingaapi.model.Servidor;
-
 import br.edu.ifrs.restinga.assinaturadigitalestagioifrsrestingaapi.model.Usuario;
 import jakarta.transaction.Transactional;
-import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 import jakarta.validation.Valid;
 import org.springframework.web.util.UriComponentsBuilder;
 
+import java.util.List;
 import java.util.Optional;
 
 
@@ -37,7 +36,7 @@ public class ServidorController extends BaseController {
     @ResponseStatus(HttpStatus.CREATED)
     public ResponseEntity salvar(@RequestBody @Valid DadosCadastroServidor dadosCadastroServidor, UriComponentsBuilder uriBuilder, @RequestHeader("Authorization") String token) {
         Servidor servidor = servidorRepository.findByUsuarioSistemaEmail(tokenService.getSubject(token.replace("Bearer ", "")));
-        if(servidor != null) {
+        if (servidor != null) {
             Optional<Curso> curso = cursoRepository.findById(dadosCadastroServidor.curso().getId());
             if (curso.isEmpty()) {
                 return servidorImplementacao.salvar(dadosCadastroServidor, null, uriBuilder);
@@ -57,23 +56,59 @@ public class ServidorController extends BaseController {
     }
 
     @PostMapping("/buscarOrientadorCurso")
-        public ResponseEntity buscarOrientadorCurso(@RequestBody Long cursoId){
+    public ResponseEntity buscarOrientadorCurso(@RequestBody Long cursoId) {
         var servidor = servidorRepository.findServidorByCurso_Id(cursoId);
+        return ResponseEntity.ok(servidor);
+    }
+
+    @GetMapping("/listaOrientadoresCursos")
+    public ResponseEntity listarOrientadores() {
+        List<Servidor> servidor = servidorRepository.findServidoresNotInEstagioOrDiretor();
         return ResponseEntity.ok(servidor);
     }
 
     @GetMapping("/excluirServidor/{id}")
     @Transactional
-    public ResponseEntity excluirServidor(@PathVariable Long id) {
-            Optional<Usuario> user = usuarioRepository.findById(id);
-            if(user.isPresent()){
-                servidorRepository.deleteServidorByUsuarioSistema(user.get());
-                usuarioRepository.deleteById(id);
-                return ResponseEntity.ok().build();
+    public ResponseEntity excluirServidor(@PathVariable Long id, @RequestHeader("Authorization") String token) {
+        String email = tokenService.getSubject(token.replace("Bearer ", ""));
+        if (!servidorRepository.existsServidorByUsuarioSistemaEmail(email)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+        Optional<Usuario> user = usuarioRepository.findById(id);
+        if (user.isPresent()) {
+            if(user.get().getRoles().getId() == 3){
+                int x = servidorRepository.countByRole_Id(user.get().getRoles().getId());
+                if(x == 1){
+                    return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Não é possível excluir o único servidor de estágio");
+                }
             }
-            else {
-                return ResponseEntity.badRequest().build();
-            }
+            servidorRepository.deleteServidorByUsuarioSistema(user.get());
+            usuarioRepository.deleteById(id);
+            return ResponseEntity.ok().build();
+        } else {
+            return ResponseEntity.badRequest().build();
+        }
+    }
+
+    @PostMapping("/excluirServidorEmail")
+    @Transactional
+    public ResponseEntity excluirServidor(@RequestBody String email, @RequestHeader("Authorization") String token) {
+        String emailToken = tokenService.getSubject(token.replace("Bearer ", ""));
+        if (email.equals(emailToken)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Não é possível excluir o próprio usuário");
+        }
+        if(!servidorRepository.existsServidorByUsuarioSistemaEmail(emailToken)){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+        Usuario user = usuarioRepository.findUsuarioByEmail(email);
+        if (user != null) {
+            alunoRepository.deleteByUsuarioSistemaEmail(email);
+            servidorRepository.deleteServidorByUsuarioSistema(user);
+            usuarioRepository.deleteById(user.getId());
+            return ResponseEntity.ok().build();
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
     @GetMapping("/listarServidores")
@@ -92,7 +127,6 @@ public class ServidorController extends BaseController {
     @GetMapping("/buscarServidoresPorEmail/{email}")
     public ResponseEntity buscarServidorPorEmail(@PathVariable String email) {
         var servidores = servidorImplementacao.listar();
-        // Percorrer a lista de servidores e encontrar o servidor com o email correspondente
         Servidor servidorEncontrado = null;
         for (Servidor servidor : servidores) {
             if (servidor.getUsuarioSistema().getEmail().equals(email)) {
